tested sl-classification on mississippi$

+++ src/bwt.c
@@ -5,7 +5,7 @@
 
 #include "cstr.h"
 
-uint8_t* cstr_bwt(int n, uint8_t const* x, unsigned int sa[n])
+uint8_t* cstr_bwt(long long n, uint8_t const* x, unsigned int sa[n])
 {
     uint8_t* b = malloc((size_t)(n + 1) * sizeof *b);
     for (int i = 0; i < n; i++) {
@@ -14,7 +14,7 @@ uint8_t* cstr_bwt(int n, uint8_t const* x, unsigned int sa[n])
     return b;
 }
 
-struct cstr_bwt_c_table* cstr_compute_bwt_c_table(int n, uint8_t const* x, int asize)
+struct cstr_bwt_c_table* cstr_compute_bwt_c_table(long long n, uint8_t const* x, int asize)
 {
     struct cstr_bwt_c_table* ctab = malloc(offsetof(struct cstr_bwt_c_table, cumsum) + (size_t)asize * sizeof(*ctab->cumsum));
     ctab->asize = asize;
@@ -41,14 +41,15 @@ void cstr_print_bwt_c_table(struct cstr_bwt_c_table const* ctab)
 
 struct cstr_bwt_o_table {
     int asize;
-    int n;
-    int table[];
+    long long n;
+    long long table[];
 };
 
 #define OTAB(otab, a, i) ((otab)->table[(i) * (otab)->asize + (a)])
 
 struct cstr_bwt_o_table*
-cstr_compute_bwt_o_table(int n, uint8_t const* bwt, struct cstr_bwt_c_table const* ctab)
+cstr_compute_bwt_o_table(long long n, uint8_t const* bwt,
+                         struct cstr_bwt_c_table const* ctab)
 {
     struct cstr_bwt_o_table* otab = malloc(offsetof(struct cstr_bwt_o_table, table) + (size_t)(ctab->asize * n) * sizeof(*otab->table));
     otab->asize = ctab->asize;
@@ -66,13 +67,13 @@ void cstr_print_bwt_o_table(struct cstr_bwt_o_table const* otab)
 {
     for (int a = 0; a < otab->asize; a++) {
         for (int i = 0; i < otab->n; i++) {
-            printf("%d ", OTAB(otab, a, i));
+            printf("%lld ", OTAB(otab, a, i));
         }
         printf("\n");
     }
 }
 
-int cstr_bwt_o_tab_rank(struct cstr_bwt_o_table const* otab, uint8_t a, int i)
+long long cstr_bwt_o_tab_rank(struct cstr_bwt_o_table const* otab, uint8_t a, long long i)
 {
     if (i == 0)
         return 0;
@@ -81,7 +82,7 @@ int cstr_bwt_o_tab_rank(struct cstr_bwt_o_table const* otab, uint8_t a, int i)
 
 // FIXME: p should be a slice!
 void cstr_bwt_search(
-    int* left, int* right,
+    long long* left, long long* right,
     uint8_t const* x, uint8_t const* p,
     struct cstr_bwt_c_table const* ctab,
     struct cstr_bwt_o_table const* otab)
